// src/features/students/types/student.types.ts
export interface StudentTypes {
    id: number;
    codigo: string;
    nombre: string;
    idGenero: string;
    genero: string;
    idEtnia: string;
    etnia: string;
    idArea: string;
    area: string;
    idRol: string;
    rol: string;
    idCalendario: string;
    calendario: string;
    idSede: string;
    sede: string;
}